% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_palettes.R
\docType{data}
\name{style_colors}
\alias{style_colors}
\title{Available Color Palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
style_colors
}
\description{
Use \code{\link{style_palette}} to extract palettes of the desired length
}
\keyword{datasets}
