% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_palettes.R
\name{style_ggplot2_scales}
\alias{style_ggplot2_scales}
\alias{scale_color_style}
\alias{scale_fill_style}
\alias{scale_color_style_c}
\alias{scale_fill_style_c}
\title{Style ggplot2 color scales}
\usage{
scale_color_style(option = "blue_gradient", direction = 1, ...)

scale_fill_style(option = "blue_gradient", direction = 1, ...)

scale_color_style_c(option = "blue_gradient", direction = 1, ...)

scale_fill_style_c(option = "blue_gradient", direction = 1, ...)
}
\arguments{
\item{option}{The name of the color palette to use. See \code{\link{style_palette}} for available options}

\item{direction}{If greater than or equal to 0, it will return the standard direction; otherwise it will reverse the direction of the palette}

\item{...}{Additional arguments passed to ggplot2 scale functions}
}
\description{
ggplot2 fill and color scales derived from the style guide

- `scale_color_style()` and `scale_fill_style()` provide discrete fill functions
- `scale_color_style_c()` and `scale_fill_style_c()` provide continuous fill functions
}
\examples{
library(ggplot2)
tmp <- data.frame(x = rnorm(100), y = rnorm(100), z = rep(c("a", "b", "c", "d", "e"), 20))

# using a discrete scale
ggplot(tmp, aes(x, y, color = z)) +
geom_point() +
scale_color_style()

# changing the palette used when creating a discrete scale
ggplot(tmp, aes(x, y, color = z)) +
geom_point() +
scale_color_style(option = "warm_spectrum")

}
