% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_palettes.R
\name{style_palette}
\alias{style_palette}
\title{Style Color Palette Constructor}
\usage{
style_palette(option = "blue_gradient", n = NULL)
}
\arguments{
\item{option}{The name of the color palette to use. Options are 'blue_gradient', 'warm_spectrum', 'cool_calm', 'vibrant_mix', and 'pastel'}

\item{n}{The number of colors to return. Currently, all scales are limited to 5 colors}
}
\description{
Extract a color palette of the desired length
}
